pyproject.toml
[project]
name = "jdhe-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.191.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "google-genai>=1.66.0",
    "pymupdf>=1.27.1",
    "python-dotenv>=1.2.2",
]
